Move exception classes to exception package
