+++ core/src/main/java/bisq/core/btc/exceptions/TxBroadcastException.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.btc.wallet;
+package bisq.core.btc.exceptions;
 
 import lombok.Getter;
