+++ core/src/main/java/bisq/core/trade/protocol/tasks/buyer_as_taker/BuyerAsTakerSignAndPublishDepositTx.java
@@ -19,8 +19,8 @@ package bisq.core.trade.protocol.tasks.buyer_as_taker;
 
 import bisq.core.btc.AddressEntry;
 import bisq.core.btc.data.RawTransactionInput;
+import bisq.core.btc.exceptions.TxBroadcastException;
 import bisq.core.btc.wallet.BtcWalletService;
-import bisq.core.btc.wallet.TxBroadcastException;
 import bisq.core.btc.wallet.TxBroadcaster;
 import bisq.core.trade.Trade;
 import bisq.core.trade.protocol.TradingPeer;
