+++ core/src/main/java/bisq/core/trade/protocol/tasks/taker/CreateTakerFeeTx.java
@@ -19,10 +19,10 @@ package bisq.core.trade.protocol.tasks.taker;
 
 import bisq.core.arbitration.Arbitrator;
 import bisq.core.btc.AddressEntry;
+import bisq.core.btc.exceptions.TxBroadcastException;
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.TradeWalletService;
-import bisq.core.btc.wallet.TxBroadcastException;
 import bisq.core.btc.wallet.TxBroadcaster;
 import bisq.core.btc.wallet.WalletService;
 import bisq.core.trade.Trade;
