+++ desktop/src/main/java/bisq/desktop/main/dao/governance/proposals/ProposalsView.java
@@ -36,7 +36,7 @@ import bisq.desktop.util.validation.BsqValidator;
 
 import bisq.core.btc.exceptions.TransactionVerificationException;
 import bisq.core.btc.exceptions.WalletException;
-import bisq.core.btc.wallet.BsqBalanceListener;
+import bisq.core.btc.listeners.BsqBalanceListener;
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.dao.DaoFacade;
 import bisq.core.dao.governance.ballot.Ballot;
