+++ desktop/src/main/java/bisq/desktop/main/offer/MutableOfferDataModel.java
@@ -22,7 +22,7 @@ import bisq.core.arbitration.Arbitrator;
 import bisq.core.btc.AddressEntry;
 import bisq.core.btc.Restrictions;
 import bisq.core.btc.listeners.BalanceListener;
-import bisq.core.btc.wallet.BsqBalanceListener;
+import bisq.core.btc.listeners.BsqBalanceListener;
 import bisq.core.btc.wallet.BsqWalletService;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.TradeWalletService;
